# pkg/docs/swagger.yaml
basePath: /api/v1
definitions:
  gateway.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  gateway.GpuMetricResponse:
    properties:
      container:
        type: string
      device:
        type: string
      gpu_id:
        type: string
      hostname:
        type: string
      labels_raw:
        type: string
      metric_name:
        type: string
      model_name:
        type: string
      namespace:
        type: string
      pod:
        type: string
      timestamp:
        type: string
      value:
        type: number
    type: object
  gateway.PaginationMeta:
    properties:
      has_next:
        type: boolean
      limit:
        type: integer
      next_cursor:
        type: string
      offset:
        type: integer
      total:
        type: integer
    type: object
  gateway.TelemetryPage:
    properties:
      data:
        items:
          $ref: '#/definitions/gateway.GpuMetricResponse'
        type: array
      pagination:
        $ref: '#/definitions/gateway.PaginationMeta'
    type: object
host: localhost:8080
info:
  contact: {}
  description: |-
    Read API for the Vantage elastic GPU telemetry pipeline.
    Exposes sorted GPU IDs and time-series metric rows from PostgreSQL.
  title: Vantage GPU Telemetry API
  version: "1.0"
paths:
  /gpus:
    get:
      description: Returns the unique list of GPU UUIDs that have telemetry data.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gateway.ErrorResponse'
      summary: List GPU IDs
      tags:
      - gpus
  /gpus/{id}/telemetry:
    get:
      description: |-
        Returns time-series metric rows for a GPU ordered newest-first (API-02).
        Optional ?start_time and/or ?end_time (RFC3339) filter the window (API-03, OQ-3).
        Use limit and offset for pagination; result is wrapped in a TelemetryPage envelope (API-05).
        In offset mode pagination.total carries the total row count for the filter; pagination.has_next signals more pages.
        Prefer cursor pagination for deep walks: pass pagination.next_cursor back as ?cursor — O(page) cost,
        stable under live ingest, no total (mutually exclusive with offset).
      parameters:
      - description: GPU UUID
        in: path
        name: id
        required: true
        type: string
      - description: Inclusive lower bound (RFC3339); omit for unbounded
        in: query
        name: start_time
        type: string
      - description: Inclusive upper bound (RFC3339); omit for unbounded
        in: query
        name: end_time
        type: string
      - description: 'Max rows to return (default/ceiling: VANTAGE_GATEWAY_MAX_ROWS)'
        in: query
        minimum: 1
        name: limit
        type: integer
      - description: 'Row offset for pagination (default: 0; mutually exclusive with
          cursor)'
        in: query
        minimum: 0
        name: offset
        type: integer
      - description: Opaque keyset cursor from a previous page's pagination.next_cursor
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gateway.TelemetryPage'
        "400":
          description: malformed start_time, end_time, limit, or offset
          schema:
            $ref: '#/definitions/gateway.ErrorResponse'
        "404":
          description: gpu_id not found
          schema:
            $ref: '#/definitions/gateway.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gateway.ErrorResponse'
      summary: Get GPU telemetry
      tags:
      - gpus
schemes:
- http
swagger: "2.0"